Improve huge_patch test (#60)
* Measure just the performance of the parser

We don't want to measure time of test data preparation because it could
be slow in some python interpreters, depending on how string
concatenation is implemented.

* Improve memory efficiency for test data preparation

Data preparation for huge_patch test could be very slow because strings
are immutable, each concatenation creates a new string and discards
the old ones.

New approach of data preparation is to concatenate large string from
smaller parts with ''.join() method. This method reduces memory usage
and enhances performance, because it minimizes the number of new string
objects created.